+++ packages/core/styles/src/packages/atomic/slider/index.ts
@@ -1,0 +1,10 @@
+import * as styles from "./nimbus-slider.css";
+
+export const slider = {
+  classnames: { ...styles },
+  vars: {
+    fillLeft: styles.fillLeftVar,
+    fillRight: styles.fillRightVar,
+    thumbLeft: styles.thumbLeftVar,
+  },
+};