+++ packages/react/src/composite/Table/src/contexts/TableContext/index.ts
@@ -1,3 +1,2 @@
 export { TableContext, useTableContext } from "./TableContext";
 export type { TableContextValue } from "./tableContext.types";
-